[CoroutineAccessors] Rehome runtime functions.
Put them where they're meant to be, in swiftCore.
