+++ include/swift/Runtime/Coro.h
@@ -21,6 +21,12 @@
 #include "swift/Runtime/Config.h"
 #include <cstddef>
 
-namespace swift {} // end namespace swift
+namespace swift {
+SWIFT_RUNTIME_EXPORT
+SWIFT_CC(swift) void *swift_coro_alloc(CoroAllocator *allocator, size_t size);
+
+SWIFT_RUNTIME_EXPORT
+SWIFT_CC(swift) void swift_coro_dealloc(CoroAllocator *allocator, void *ptr);
+} // end namespace swift
 
 #endif
