+++ stdlib/public/runtime/Coro.cpp
@@ -15,3 +15,19 @@
 #include "swift/Basic/FlagSet.h"
 
 using namespace swift;
+
+void *swift::swift_coro_alloc(CoroAllocator *allocator, size_t size) {
+  return allocator->allocate(size);
+}
+
+void swift::swift_coro_dealloc(CoroAllocator *allocator, void *ptr) {
+  if (!allocator)
+    return;
+  // Calls to swift_coro_dealloc are emitted in resume funclets for every
+  // live-across dynamic allocation.  Whether such calls immediately deallocate
+  // memory depends on the allocator.
+  if (!allocator->shouldDeallocateImmediately()) {
+    return;
+  }
+  allocator->deallocate(ptr);
+}
